requirements.txt
torch==2.5.1
mysqlclient == 2.2.7
transformers==4.40.0
django>=4.2
langchain-community==0.3.23
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.2
llama-cpp-python>=0.2.28
numpy>=1.24.3
faiss-cpu>=1.7.4
regex>=2023.10.3
gunicorn>=20.1.0